% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxDtLaunchByOrganisation}
\alias{boxDtLaunchByOrganisation}
\title{boxDtLaunchByOrganisation(): Druckt einen Boxplot der Ausrückzeiten für alle
Stützpunkte in simData$vehicles}
\usage{
boxDtLaunchByOrganisation(simMissions)
}
\arguments{
\item{simScenario}{A simScenario}
}
\value{
ok
}
\description{
boxDtLaunchByOrganisation(): Druckt einen Boxplot der Ausrückzeiten für alle
Stützpunkte in simData$vehicles
}

