% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxDtLaunchByVehicleName}
\alias{boxDtLaunchByVehicleName}
\title{boxDtLaunchByVehicleName(): Druckt einen Boxplot der Ausrückzeiten für alle
Stützpunkte in simData$vehicles}
\usage{
boxDtLaunchByVehicleName(simScenario)
}
\arguments{
\item{simScenario}{A simScenario}
}
\value{
ok
}
\description{
boxDtLaunchByVehicleName(): Druckt einen Boxplot der Ausrückzeiten für alle
Stützpunkte in simData$vehicles
}

