% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{cpDtService}
\alias{cpDtService}
\title{cpDtService: color palette to paint dots according to their level of service
time (DtService).}
\usage{
cpDtService(x)
}
\arguments{
\item{x}{A number between 0 and 100 (percentage)}
}
\value{
color A string containing the name of a color
}
\description{
cpDtService: color palette to paint dots according to their level of service
time (DtService).
}

