% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{fhsblue}
\alias{fhsblue}
\title{fhsblue: brings you the funky fhsblue and makes all graphics corporate design
compatible.}
\usage{
fhsblue()
}
\value{
fhsblue The famous color of all colors
}
\description{
fhsblue: brings you the funky fhsblue and makes all graphics corporate design
compatible.
}

