% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{filterPrio1}
\alias{filterPrio1}
\title{filterPrio1: Provisorische Funktion für das Filtern aller Events mit
Priorität 1. Achtung: Die Logik beim rausfiltern ist nicht in allen
Teilen von plot11R konsistent gelöst. Teilweise geschieht die Filterung
explizit vor dem Funktionisaufruf, teilweise implizit in der Funktion}
\usage{
filterPrio1(simData)
}
\arguments{
\item{simData}{A set of simMissions or simEvents}
}
\value{
simEventsPrio12 A subset of simEvents with events of priority 1 & 2
}
\description{
filterPrio1: Provisorische Funktion für das Filtern aller Events mit
Priorität 1. Achtung: Die Logik beim rausfiltern ist nicht in allen
Teilen von plot11R konsistent gelöst. Teilweise geschieht die Filterung
explizit vor dem Funktionisaufruf, teilweise implizit in der Funktion
}

