% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{filterPrio12}
\alias{filterPrio12}
\title{filterPrio12: Provisorische Funktion für das Filtern aller Events mit
Priorität 1 und 2. Achtung: Die Logik beim rausfiltern ist nicht in allen
Teilen von plot11R konsistent gelöst. Teilweise geschieht die Filterung
explizit vor dem Funktionisaufruf, teilweise implizit in der Funktion}
\usage{
filterPrio12(simEvents)
}
\arguments{
\item{simEvents}{}
}
\value{
simEventsPrio12 A subset of simEvents with events of priority 1 & 2
}
\description{
filterPrio12: Provisorische Funktion für das Filtern aller Events mit
Priorität 1 und 2. Achtung: Die Logik beim rausfiltern ist nicht in allen
Teilen von plot11R konsistent gelöst. Teilweise geschieht die Filterung
explizit vor dem Funktionisaufruf, teilweise implizit in der Funktion
}

