% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/histograms.R
\name{histDtService}
\alias{histDtService}
\title{histDtService: that famous (but not very useful) histogram of all events.}
\usage{
histDtService(simEvents)
}
\arguments{
\item{simEvents}{A set of simEvents}
}
\value{
plotted histogram
}
\description{
histDtService: that famous (but not very useful) histogram of all events.
}

