% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/histograms.R
\name{histTAlarmHour}
\alias{histTAlarmHour}
\title{histTAlarmHour:}
\usage{
histTAlarmHour(simMissions)
}
\arguments{
\item{simEvents}{A set of simMissions}
}
\value{
plotted histogram
}
\description{
histTAlarmHour:
}

