% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/histograms.R
\name{histTAlarmMonth}
\alias{histTAlarmMonth}
\title{histTAlarmMonth:}
\usage{
histTAlarmMonth(simMissions)
}
\arguments{
\item{simMissions}{A set of simMissions}
}
\value{
plotted histogram
}
\description{
histTAlarmMonth:
}

