% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kpis.R
\name{kpiEventsFaster}
\alias{kpiEventsFaster}
\title{kpiEventsFaster(): Anzahl der Events, welche schneller als im
zugrundeliegenden Referenzszenario erreicht wurden.}
\usage{
kpiEventsFaster(simEvents, simEventsRef)
}
\arguments{
\item{simEvents}{}

\item{simEventsRef}{}
}
\value{
kpiEventsFaster
}
\description{
kpiEventsFaster(): Anzahl der Events, welche schneller als im
zugrundeliegenden Referenzszenario erreicht wurden.
}

