% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kpis.R
\name{kpiEventsFasterPercentage}
\alias{kpiEventsFasterPercentage}
\title{kpiEventsFasterPercentage(): Prozentsatz der Events, welche schneller als
im zugrundeliegenden Referenzszenario erreicht wurden.}
\usage{
kpiEventsFasterPercentage(simEvents, simEventsRef)
}
\arguments{
\item{simEvents}{}

\item{simEventsRef}{}
}
\value{
kpiEventsFasterPercentage A number
}
\description{
kpiEventsFasterPercentage(): Prozentsatz der Events, welche schneller als
im zugrundeliegenden Referenzszenario erreicht wurden.
}

