% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kpis.R
\name{kpiEventsInTime}
\alias{kpiEventsInTime}
\title{kpiEventsInTime: Calculates the percentage of events of priority 1&2 which
have a dtService of less than 15min. In other words: which are in time.
Rounded to 1 decimalplace.}
\usage{
kpiEventsInTime(simEvents)
}
\arguments{
\item{simEvents}{A set of simEvents}
}
\value{
EventsInTime A percentage of the fraction of the events in time
}
\description{
kpiEventsInTime: Calculates the percentage of events of priority 1&2 which
have a dtService of less than 15min. In other words: which are in time.
Rounded to 1 decimalplace.
}

