% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kpis.R
\name{kpiEventsInTimeDelta}
\alias{kpiEventsInTimeDelta}
\title{kpiEventsInTimeDelta: Calculates the delta of the percentage of the events
in time between two sets of simEvents. Rounded to 1 decimalplace.}
\usage{
kpiEventsInTimeDelta(simEvents, simEventsBefore)
}
\arguments{
\item{simEvents}{}

\item{simEventsBefore}{}
}
\value{
EventsInTimeDelta A numeric
}
\description{
kpiEventsInTimeDelta: Calculates the delta of the percentage of the events
in time between two sets of simEvents. Rounded to 1 decimalplace.
}

