% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kpis.R
\name{kpiEventsNewInTime}
\alias{kpiEventsNewInTime}
\title{kpiEventsNewInTime(): Errechnet die Anzahl der Ereignisse, die neu innerhalb
15 min. erreicht werden.}
\usage{
kpiEventsNewInTime(simEvents, simEventsBefore)
}
\arguments{
\item{simEvents}{}

\item{simEventsBefore}{}
}
\value{
numeric
}
\description{
kpiEventsNewInTime(): Errechnet die Anzahl der Ereignisse, die neu innerhalb
15 min. erreicht werden.
}

