% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kpis.R
\name{kpiMissionsExternalServiceNeeded}
\alias{kpiMissionsExternalServiceNeeded}
\title{kpiMissionsExternalServiceNeeded(): Returns the number of missions that used
the external vehicles as specified in vehicleIdExt}
\usage{
kpiMissionsExternalServiceNeeded(simMissions, vehicleIdExt)
}
\arguments{
\item{simMissions}{A set of simMissions}

\item{vehicleIdExt}{A (vector) of vehicleId's used by external services.}
}
\value{
nrMissionsExternalServiceNeeded A numeric
}
\description{
kpiMissionsExternalServiceNeeded(): Returns the number of missions that used
the external vehicles as specified in vehicleIdExt
}

