% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kpis.R
\name{kpiMissionsInTime}
\alias{kpiMissionsInTime}
\title{kpiMissionsInTime: Calculates the percentage of events which
have a dtService of less than 15min. In other words: which are in time.}
\usage{
kpiMissionsInTime(simMissions)
}
\arguments{
\item{simMissions}{A set of simMissions}
}
\value{
MissionsInTime A percentage of the fraction of the events in time
}
\description{
kpiMissionsInTime: Calculates the percentage of events which
have a dtService of less than 15min. In other words: which are in time.
}

