% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kpis.R
\name{kpiMissionsInTimeDelta}
\alias{kpiMissionsInTimeDelta}
\title{kpiMissionsInTimeDelta: Calculates the delta of the percentage of the missions
in time between two sets of simMissions. Rounded to 1 decimalplace.}
\usage{
kpiMissionsInTimeDelta(simMissions, simMissionsRef)
}
\arguments{
\item{simMissions}{A set of simMissions}

\item{simMissionsRef}{A set of simMissions}
}
\value{
MissionsInTimeDelta A numeric
}
\description{
kpiMissionsInTimeDelta: Calculates the delta of the percentage of the missions
in time between two sets of simMissions. Rounded to 1 decimalplace.
}

