% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maps.R
\name{mapDtServicePoA}
\alias{mapDtServicePoA}
\title{mapDtServicePoA: Die famose Karte aufgelöst nach Einsatzort (für Bern
bedeutet das aufgelöst nach Ort)}
\usage{
mapDtServicePoA(simMissions)
}
\arguments{
\item{simMissions}{}
}
\value{
mapDtServicePoA A leaflet map
}
\description{
mapDtServicePoA: Die famose Karte aufgelöst nach Einsatzort (für Bern
bedeutet das aufgelöst nach Ort)
}

