% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maps.R
\name{mapDtToPoA}
\alias{mapDtToPoA}
\title{mapDtToPoA(): Die blau/gelb Karte aufgelöst nach PoA}
\usage{
mapDtToPoA(simMissions, simMissionsRef)
}
\arguments{
\item{simMissions}{A set of simMissions}

\item{simMissionsRef}{A set of simMissions}
}
\value{
mapDtToPoA A leaflet map
}
\description{
mapDtToPoA(): Die blau/gelb Karte aufgelöst nach PoA
}

