% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maps.R
\name{mapOrganisationPrio12}
\alias{mapOrganisationPrio12}
\title{mapOrganisationPrio12(): Die famose Karte mit einem Punkt für jeden
Rettungsdienst. Die Farbe zeigt die Hilfsfristeinhaltung.}
\usage{
mapOrganisationPrio12(simScenario)
}
\arguments{
\item{simScenario}{}
}
\value{
map A leaflet object
}
\description{
mapOrganisationPrio12(): Die famose Karte mit einem Punkt für jeden
Rettungsdienst. Die Farbe zeigt die Hilfsfristeinhaltung.
}

