% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maps.R
\name{mapPoAPrio12}
\alias{mapPoAPrio12}
\title{mapPoAPrio12: Die famose Karte aufgelöst nach Einsatzort (für Bern
bedeutet das aufgelöst nach Ort)}
\usage{
mapPoAPrio12(simMissions)
}
\arguments{
\item{simMissions}{}
}
\value{
mapPoAPrio12 A leaflet map
}
\description{
mapPoAPrio12: Die famose Karte aufgelöst nach Einsatzort (für Bern
bedeutet das aufgelöst nach Ort)
}

