% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables.R
\name{tableDtService}
\alias{tableDtService}
\title{tableDtService(): Liefert eine Tabelle mit ausgesuchten Spalten zweier Sets
von simMissions}
\usage{
tableDtService(simScenario, simScenarioRef)
}
\arguments{
\item{simMissions}{A set of simMissions}

\item{simMissionsBefore}{A set of simMissions}
}
\value{
tableDtService A data.frame
}
\description{
tableDtService(): Liefert eine Tabelle mit ausgesuchten Spalten zweier Sets
von simMissions
}

