% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables.R
\name{tableDtServicePrio12}
\alias{tableDtServicePrio12}
\title{tableDtServicePrio12}
\usage{
tableDtServicePrio12(simScenario, simScenarioRef)
}
\arguments{
\item{simScenario}{}

\item{simScenarioRef}{}
}
\value{
tableDtServicePrio12 A data.frame
}
\description{
tableDtServicePrio12
}

