% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables.R
\name{tableMissionsNewInTime}
\alias{tableMissionsNewInTime}
\title{tableMissionsNewInTime(): Alle Einsätze, die neu innerhalb der Frist erreicht
werden als Tabelle}
\usage{
tableMissionsNewInTime(simMissions, simMissionsBefore)
}
\arguments{
\item{simMissions}{A set of simMissions}

\item{simMissionsBefore}{A set of simMissions}
}
\value{
simMissionsNewInTime A set of simMissions
}
\description{
tableMissionsNewInTime(): Alle Einsätze, die neu innerhalb der Frist erreicht
werden als Tabelle
}

