% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables.R
\name{tableNrOfMissionsByOrganisation}
\alias{tableNrOfMissionsByOrganisation}
\title{tableNrOfMissionsByOrganisation(): Tabelle mit den explizit gemachten
Verschiebungen zwischen den Rettungsdiensten.}
\usage{
tableNrOfMissionsByOrganisation(simScenario, simScenarioRef)
}
\arguments{
\item{simScenario}{A simScenario}

\item{simScenarioRef}{A simScenario}
}
\value{
tableNrOfMissionsByOrganisation A data.frame
}
\description{
tableNrOfMissionsByOrganisation(): Tabelle mit den explizit gemachten
Verschiebungen zwischen den Rettungsdiensten.
}

